Create Tokenziation
we want the textual data to be tokenized. NLTK provides tokenizer to tokens with some range.  Tokens means splitting of data into words in some order. once tokenizer implemented in the reviews then the next form is about to  convert text sequence into numbers which helps the system to convert text data into numerical one. 
Sequence of data looks like a sparse matrix which is inefficient to process on any models so we need to pad the sequence which means make the dimension proper for the given data