Create neualnetwork.py
Here LSTM, Neural Network Model created and for future use the model dimensions were saving to disk.
This is how the model has been designed here, along with that we are storing the model weights into the disk by storing in the json file for future purpose. While storing this model we can utilize for transfer learning and other purpose. Again we load the data from the json file and checking the accuracy rate. While using this type of mechanism. Each time the model runs it store the accuracy level and keep increasing the factor. That’s how the system works.