Create segmentation.py
Segmentation of Words are processed here, finding of Positive Expressions and Negative Expressions were labeled here.
With the bigram finders we apply set of conditions. We use NLTK library to extract bigram collocation. Bigram - choice of two words. The expression consisting of two words that respond in a conventional way. In this project we concentrate to build the bigram for that initially we separate the reviews into a list of two words after that we are finding the frequency for the same bigram word and also  we store it in variable.